+++ docs/changes.rst
@@ -11,6 +11,7 @@ Next version
 * Fixed ``RENDER_PANELS`` functionality so that when ``True`` panels are
   rendered during the request and not loaded asynchronously.
 * HistoryPanel now shows status codes of responses.
+* Support ``request.urlconf`` override when checking for toolbar requests.
 
 
 3.2.1 (2021-04-14)